refactor(chat): extract media + attachment handlers from CLVM
`ConversationListViewModel.kt` was 3564 lines; this PR drops it to 2625 by
moving two cohesive action-arm clusters into sibling handler files in the
same package, with no public-API change.

The pre-paging plan originally framed this work as "split the VM in two"
(list-of-conversations vs open-conversation). After reading the file the
split-by-VM idea was downgraded — the two halves share `_uiState.ownerSession`,
`_uiState.activeConversations`, `_uiState.selectedConversationId` and the
notification-tap pipeline; a literal class split would force those reads
through a coordinator/repo invented just for the split, and re-key all 41
`viewModel.*` references in `ConversationListScreen.kt`. Handler extraction
keeps the public API byte-identical and is mechanical to review.

What moved
- `MediaDownloadHandler.kt` (new, ~545 lines): 9 action arms
  ShareMedia, ShareMessage, DownloadMedia, DownloadVideoMedia, DecryptFile,
  SaveFile, MediaClicked, ShowMoreClicked, CloseFullScreenOverlay
  + 3 private helpers
  resolveHlsVideoMetadata, downloadAndRemuxHlsToMp4, resolveDownloadFileName.
- `AttachmentHandler.kt` (new, ~557 lines): 13 action arms
  AttachPlatformFile, AttachGalleryItem, UnAttachFile, AttachClipboardImage,
  RequestCropAttachment, ApplyCropResult, RequestDrawAttachment, ApplyDrawResult,
  ApplyTrimResult, ShowRecordingHelp, StartRecording, StopRecording,
  CancelRecording
  + the `pendingThumbnails` map and `extractThumbnailAsync` / `ensureThumbnail`
  helpers (the latter exposed as `internal` so the VM's still-resident
  `addMessageWithFiles` can await poster frames before sending).

Both handlers are constructed in the VM body via property initializers and
hold function references to `::onAction`, `::sendEvent`, and (for the
attachment handler) `::addMessageWithFiles`. Each is `internal class` in
the same package so `MutableStateFlow` access stays encapsulated to the
module.

What was preserved verbatim
- Every action arm body is a verbatim move; the only renames are
  `viewModelScope` -> `scope`, `_uiState` -> `uiState`,
  `_messagesUiState` -> `messagesUiState`, `onAction(...)` -> `dispatch(...)`.
  Confirmed by grep: zero `viewModelScope|_uiState|_messagesUiState`
  references remain in either handler file.
- `DecryptFile`'s pre-existing dual-state mutation (writes to
  `_messagesUiState.decryptedFiles` on success, removes from
  `_uiState.downloadingFiles` in `finally` — the entry add is missing on
  main and may be dead code) is preserved unchanged. Worth investigating
  in a follow-up; not in scope here.
- The unicode-NUL-character literal sanitizer in `resolveDownloadFileName`
  / `downloadAndRemuxHlsToMp4` preserved (had to patch via Python after
  the Write tool turned the 6-char Kotlin escape sequence into a literal
  NUL byte in the source file).

What changed in the VM
- Replaced 22 action-arm bodies with single-line delegate calls to the
  handlers.
- Deleted 5 private helpers and the `pendingThumbnails` field.
- Rewired the one remaining call site of `ensureThumbnail` (in
  `addMessageWithFiles`) to `attachmentHandler.ensureThumbnail(f)`.
- Dropped 24 imports that were only used by the moved code (caught one
  false-positive — `io.github.vinceglb.filekit.mimeType` was still used
  by `addMessageWithFiles` — and restored it on the next compile).

Verification
- Compiles clean on all 6 targets:
  `homebase-chat:compileKotlinJvm`, `compileAndroidMain`,
  `compileKotlinIosSimulatorArm64`, `homebase-core:compileKotlinJvm`,
  `androidApp:compileDebugKotlin`, `desktopApp:compileKotlinJvm`.
- `homebase-chat:jvmTest --rerun-tasks` green.
- The `homebase-core:jvmTest` `NotificationTapColdStartTest` flake
  reproduces on clean main (SkikoComposeUiTest lifecycle teardown
  `IllegalStateException`); not introduced here.

Not changed
- Public VM API: `uiState`, `messagesUiState`, `onAction`, `eventConsumed`,
  `dialogClosed`, `conversationSearchTextState`, `messageInputTextState`,
  `messagesSearchTextState` — all byte-identical.
- `ConversationListScreen.kt` (41 `viewModel.*` references), Koin module,
  nav graph, cross-module callers.
- The 23-param VM constructor.
- The remaining ~75% of action arms (message-actions, conversation-
  lifecycle, forwarding, search/nav glue). Those are PR 4b parking-lot
  candidates.
